requirements_django.txt
django
djangorestframework==3.14.0
pandas==2.0.3
numpy==1.24.3
django-cors-headers==4.3.1
streamlit==1.28.1
plotly==5.17.0
requests==2.31.0